refactor : request, response dto 분기
