+++ src/main/java/inu/codin/codin/domain/chat/chatroom/dto/request/ChatRoomCreateRequestDto.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.chat.chatroom.dto;
+package inu.codin.codin.domain.chat.chatroom.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
