+++ src/main/java/inu/codin/codin/domain/chat/chatroom/entity/ChatRoom.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.domain.chat.chatroom.entity;
 
 import inu.codin.codin.common.dto.BaseTimeEntity;
-import inu.codin.codin.domain.chat.chatroom.dto.ChatRoomCreateRequestDto;
+import inu.codin.codin.domain.chat.chatroom.dto.request.ChatRoomCreateRequestDto;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
 import lombok.Builder;
