+++ src/main/java/inu/codin/codin/domain/email/dto/request/JoinEmailSendRequestDto.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.email.dto;
+package inu.codin.codin.domain.email.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
