+++ src/main/java/inu/codin/codin/domain/email/service/EmailAuthService.java
@@ -1,8 +1,8 @@
 package inu.codin.codin.domain.email.service;
 
 import inu.codin.codin.common.exception.NotFoundException;
-import inu.codin.codin.domain.email.dto.JoinEmailCheckRequestDto;
-import inu.codin.codin.domain.email.dto.JoinEmailSendRequestDto;
+import inu.codin.codin.domain.email.dto.request.JoinEmailCheckRequestDto;
+import inu.codin.codin.domain.email.dto.request.JoinEmailSendRequestDto;
 import inu.codin.codin.domain.email.entity.EmailAuthEntity;
 import inu.codin.codin.domain.email.exception.EmailAuthFailException;
 import inu.codin.codin.domain.email.repository.EmailAuthRepository;
