+++ src/main/java/inu/codin/codin/domain/lecture/domain/review/controller/ReviewController.java
@@ -2,7 +2,7 @@ package inu.codin.codin.domain.lecture.domain.review.controller;
 
 import inu.codin.codin.common.response.SingleResponse;
 import inu.codin.codin.domain.lecture.domain.review.service.ReviewService;
-import inu.codin.codin.domain.lecture.domain.review.dto.CreateReviewRequestDto;
+import inu.codin.codin.domain.lecture.domain.review.dto.request.CreateReviewRequestDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
