+++ src/main/java/inu/codin/codin/domain/lecture/domain/review/dto/request/CreateReviewRequestDto.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.lecture.domain.review.dto;
+package inu.codin.codin.domain.lecture.domain.review.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
