+++ src/main/java/inu/codin/codin/domain/lecture/domain/review/dto/response/ReviewListResposneDto.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.lecture.domain.review.dto;
+package inu.codin.codin.domain.lecture.domain.review.dto.response;
 
 
 import inu.codin.codin.domain.lecture.domain.review.entity.ReviewEntity;
