+++ src/main/java/inu/codin/codin/domain/lecture/domain/review/entity/ReviewEntity.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.domain.lecture.domain.review.entity;
 
 import inu.codin.codin.common.dto.BaseTimeEntity;
-import inu.codin.codin.domain.lecture.domain.review.dto.CreateReviewRequestDto;
+import inu.codin.codin.domain.lecture.domain.review.dto.request.CreateReviewRequestDto;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
