+++ src/main/java/inu/codin/codin/domain/lecture/dto/response/LectureListResponseDto.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.lecture.dto;
+package inu.codin.codin.domain.lecture.dto.response;
 
 import inu.codin.codin.domain.lecture.entity.LectureEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
