+++ src/main/java/inu/codin/codin/domain/lecture/service/LectureService.java
@@ -3,6 +3,10 @@ package inu.codin.codin.domain.lecture.service;
 import inu.codin.codin.common.dto.Department;
 import inu.codin.codin.common.exception.NotFoundException;
 import inu.codin.codin.domain.lecture.dto.*;
+import inu.codin.codin.domain.lecture.dto.response.LectureDetailResponseDto;
+import inu.codin.codin.domain.lecture.dto.response.LectureListResponseDto;
+import inu.codin.codin.domain.lecture.dto.response.LecturePageResponse;
+import inu.codin.codin.domain.lecture.dto.response.LectureSearchListResponseDto;
 import inu.codin.codin.domain.lecture.entity.LectureEntity;
 import inu.codin.codin.domain.lecture.exception.WrongInputException;
 import inu.codin.codin.domain.lecture.repository.LectureRepository;
