+++ src/main/java/inu/codin/codin/domain/notification/service/NotificationService.java
@@ -3,7 +3,7 @@ package inu.codin.codin.domain.notification.service;
 import inu.codin.codin.common.exception.NotFoundException;
 import inu.codin.codin.common.security.util.SecurityUtils;
 import inu.codin.codin.domain.like.entity.LikeType;
-import inu.codin.codin.domain.notification.dto.NotificationListResponseDto;
+import inu.codin.codin.domain.notification.dto.response.NotificationListResponseDto;
 import inu.codin.codin.domain.notification.entity.NotificationEntity;
 import inu.codin.codin.domain.notification.repository.NotificationRepository;
 import inu.codin.codin.domain.post.domain.comment.entity.CommentEntity;
