+++ src/main/java/inu/codin/codin/infra/fcm/dto/request/FcmTokenRequest.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.infra.fcm.dto;
+package inu.codin.codin.infra.fcm.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
